deprecate groups, use `!`s to convey priority
Overview
--------

For most transaction journals, the order of entries can be determined
by sorting by date. Nightscape does not enforce order of entries in its
transaction journals, however, so it is possible for the order of entries
to become jumbled if two or more entries occur on the same date but appear
in disparate locations of the transaction journal.

While grouping Entry IDs together allows for unjumbling based on a
given array index, if two Entries belong to the same two groups, it is
possible for Entry ID 1 to have higher sort priority according to the
first group and lower sort priority according to the second. Lexically
sorting groups would be required to handle these corner cases.

Groups have been deprecated in favor of exclamation marks `!`, which
convey priority sorting in a minimum of characters. As hashtags can group
entries under a common name, with the introduction of exclamation marks,
groups are superfluous.

The `!` takes up less space than groups and seems more intuitive.
The more exclamation marks an Entry has, the higher is that Entry's
priority when sorting.

Example
-------

salary, due to its one exclamation point, will appear ahead of other
entries on 2014-01-01:

```transasctions
; personal {{{

; coffee --- {{{

2014-01-01 "I bought coffee"
  Expenses:Personal:Fuel             $20.00 USD
  Assets:Personal:Bankwest:Cheque   -$20.00 USD

; ---- }}}
; fuel --- {{{

2014-01-01 "I bought fuel"
  Expenses:Personal:Fuel             $30.00 USD
  Assets:Personal:Bankwest:Cheque   -$30.00 USD

; ------ }}}
; salary --- {{{

2014-01-01 "I was paid $10000" !    ; important
  Assets:Personal:Bankwest:Cheque    $10000.00 USD
  Income:Personal:Salary             $10000.00 USD

; ------- }}}

; }}}
```